trailer-detection/sequence_detection/requirements.txt
Keras==2.2.2
matplotlib==2.2.2
numpy==1.15.1
opencv-python==3.4.0.12
pandas==0.22.0
Pillow==5.2.0
PyWavelets==0.5.2
scipy==1.1.0
tensorflow-gpu==1.9.0